+++ src/common/components/PieChart/Divider.js
@@ -1,9 +1,10 @@
 import { PropTypes } from 'react';
 
 import styles from './styles.less';
+import { prefix } from 'lib/css';
 
 const Divider = ({ rotationOffset }) => (
-  <div className={styles.divider} style={{ transform: `rotate(${rotationOffset}deg)` }} />
+  <div className={styles.divider} style={prefix('transform', `rotate(${rotationOffset}deg)`)} />
 );
 
 Divider.propTypes = {